Custom error renamed and ErrorHandler created
