+++ src/domain/errors/custom-error.ts
