+++ src/domain/index.ts
@@ -7,7 +7,8 @@ export * from './entities/genre.entity';
 export * from './entities/movie.entity';
 export * from './entities/review.entity';
 export * from './entities/user.entity';
-export * from './errors/custom.error';
+export * from './errors/custom-error';
+export * from './errors/error-handler';
 export * from './interfaces/auth/email-auth.interfaces';
 export * from './interfaces/auth/login-auth.interfaces';
 export * from './interfaces/auth/oauth-auth.interfaces';
